% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{BannadiBurials}
\alias{BannadiBurials}
\title{Bannadi Burials Data}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}

\describe{
}
}
\source{
Manly, B 1994)
}
\usage{
BannadiBurials
}
\description{
Bannadi Burials Data
}
\author{
contributed by A. Jonathan R. Godfrey
}
