% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{Belize}
\alias{Belize}
\title{Belize Soil Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 151 observations on the following eight variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{X4}}{a numeric vector}
\item{\code{Y1}}{a numeric vector}
\item{\code{Y2}}{a numeric vector}
\item{\code{Y3}}{a numeric vector}
\item{\code{Y4}}{a numeric vector}
}
}
\usage{
Belize
}
\description{
Table 10.2 on page 158  of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
