% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{BumpusSparrows}
\alias{BumpusSparrows}
\title{Bumpus Sparrows Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 49 observations on the following six variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{Survivorship}}{a character vector}
\item{\code{Total_length}}{a numeric vector}
\item{\code{Alar_extent}}{a numeric vector}
\item{\code{L_beak_head}}{a numeric vector}
\item{\code{L_humerous}}{a numeric vector}
\item{\code{L_keel_sternum}}{a numeric vector}
}
}
\usage{
BumpusSparrows
}
\description{
Table 1.1 on page 2 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
