% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{ButterflyColonies}
\alias{ButterflyColonies}
\title{Butterfly Colonies Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 16 observations on the following 10 variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{Altitude}}{a numeric vector}
\item{\code{Ann.Prec}}{a numeric vector}
\item{\code{Max.Temp}}{a numeric vector}
\item{\code{Min.Temp}}{a numeric vector}
\item{\code{PFPGi0.40}}{a numeric vector}
\item{\code{PFPGi0.60}}{a numeric vector}
\item{\code{PFPGi0.80}}{a numeric vector}
\item{\code{PFPGi1.00}}{a numeric vector}
\item{\code{PFPGi1.16}}{a numeric vector}
\item{\code{PFPGi1.30}}{a numeric vector}
}
}
\usage{
ButterflyColonies
}
\description{
Table 1.3 on page 8 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
