% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{EgyptianSkulls}
\alias{EgyptianSkulls}
\title{Egyptian Skulls Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 150 observations on the following five variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{Period}}{a character vector}
\item{\code{Maximum.breadth}}{a numeric vector}
\item{\code{Basibregmatic.height}}{a numeric vector}
\item{\code{Basialveolar.length}}{a numeric vector}
\item{\code{Nasal.height}}{a numeric vector}
}
}
\source{
also found as dataset No. 361 in Hand et al. (1994)
}
\usage{
EgyptianSkulls
}
\description{
Table 1.2 on page 4 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
