% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{EuroEmp}
\alias{EuroEmp}
\title{European Employment Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with thirty observations on the following eleven variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{Country}}{a character vector}
\item{\code{Group}}{a character vector}
\item{\code{AGR}}{a numeric vector}
\item{\code{MIN}}{a numeric vector}
\item{\code{MAN}}{a numeric vector}
\item{\code{PS}}{a numeric vector}
\item{\code{CON}}{a numeric vector}
\item{\code{SER}}{a numeric vector}
\item{\code{FIN}}{a numeric vector}
\item{\code{SPS}}{a numeric vector}
\item{\code{TC}}{a numeric vector}
}
}
\usage{
EuroEmp
}
\description{
Table 1.5 on page 11 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
