% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{EuroFood}
\alias{EuroFood}
\title{European Food Consumption Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 25 observations on the following ten variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{X}}{a character vector}
\item{\code{RedMeat}}{a numeric vector}
\item{\code{WhiteMeat}}{a numeric vector}
\item{\code{Eggs}}{a numeric vector}
\item{\code{Milk}}{a numeric vector}
\item{\code{Fish}}{a numeric vector}
\item{\code{Cereals}}{a numeric vector}
\item{\code{Starch}}{a numeric vector}
\item{\code{Nuts}}{a numeric vector}
\item{\code{FruitVeg}}{a numeric vector}
}
}
\usage{
EuroFood
}
\description{
Table 6.7 on page  90 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
