% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{Goblets}
\alias{Goblets}
\title{Goblets Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 25 observations on the following six variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{InrWth}}{a numeric vector}
\item{\code{BwlWth}}{a numeric vector}
\item{\code{TtlHt}}{a numeric vector}
\item{\code{BaseWth}}{a numeric vector}
\item{\code{StmWth}}{a numeric vector}
\item{\code{StmHt}}{a numeric vector}
}
}
\usage{
Goblets
}
\description{
Table 6.6 on page 89  of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
