% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JohnsonWichern2007Data.R
\name{JW2007}
\alias{JW2007}
\title{Data from Johnson and Wichern (2007)}
\description{
Tables from Johnson and Wichern (2007) have been imported from csv files. These were created from tab delimited text files.
}
\details{
The data has undergone minimal cleaning.
}
\references{
Johnson, and Wichern, (2007) \emph{Applied Multivariate Statistical Analysis}, sixth edition,
}
\author{
contributed by A. Jonathan R. Godfrey
}
