% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{PrehistoricDogs}
\alias{PrehistoricDogs}
\title{PrehistoricDogs Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with observations on the following variables.
}\if{html}{\out{</div>}}

\describe{
}
}
\usage{
PrehistoricDogs
}
\description{
Table 1.4 on page 10 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
