% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{Steneryd}
\alias{Steneryd}
\title{Steneryd Reserve Plants Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 25 observations on the following 17 variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{X1}}{a numeric vector}
\item{...}{...}
\item{\code{X17}}{a numeric vector}
}
}
\usage{
Steneryd
}
\description{
Table 9.3 on page 142 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
