% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{VotingNJ}
\alias{VotingNJ}
\title{Voting of New Jersey Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with 15 observations on the following 15 variables.
}\if{html}{\out{</div>}}

\describe{
\item{\code{Hunt.R}}{a numeric vector}
\item{\code{Sandman.R}}{a numeric vector}
\item{\code{Howard.D}}{a numeric vector}
\item{\code{Thompson.D}}{a numeric vector}
\item{\code{Frelinghuysen.R}}{a numeric vector}
\item{\code{Forsythe.R}}{a numeric vector}
\item{\code{Widnall.R}}{a numeric vector}
\item{\code{Roe.D}}{a numeric vector}
\item{\code{Helstoski.D}}{a numeric vector}
\item{\code{Rodino.D}}{a numeric vector}
\item{\code{Minish.D}}{a numeric vector}
\item{\code{Rinaldo.R}}{a numeric vector}
\item{\code{Maraziti.R}}{a numeric vector}
\item{\code{Daniels.D}}{a numeric vector}
\item{\code{Pattern.D}}{a numeric vector}
}
}
\usage{
VotingNJ
}
\description{
Table 11.3 on page 177 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
