% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManlyData.R
\name{RoadsNZ}
\alias{RoadsNZ}
\title{Road Distances between NZ Towns Data from Manly (1994)}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A data frame with distances between Alexandra, Balclutha, Blenheim, Christchurch, Dunedin, Franz.Josef, Greymouth, Invercargill, Milford, Nelson, Queenstown, Te.Anau, and Timaru.
}\if{html}{\out{</div>}}
}
\usage{
RoadsNZ
}
\description{
Table 11.1 on page 173 of Manly (1994)
}
\references{
Manly, B.F.J. (1994) \emph{Multivariate Statistical Methods: A Primer}, Second Edition, CRC Press.
}
\author{
contributed by A. Jonathan R. Godfrey
}
